fix(issue-433): Address PR review - harden refresh moves
