+++ cli/src/commands/refresh_sources.rs
@@ -1,32 +1,90 @@
-use crate::config;
-use std::path::PathBuf;
+use crate::agent::Agent;
+use crate::config::{self, ItemKind};
+use crate::hook::Hook;
+use crate::mapping::MappingConfig;
+use crate::pi_extension::PiExtension;
+use crate::skill::Skill;
+use std::path::{Path, PathBuf};
+
+#[derive(Clone, Debug)]
+pub(crate) struct ResolvedSource {
+    pub root: PathBuf,
+    pub aliases: Vec<String>,
+}
+
+#[derive(Clone)]
+pub struct RefreshSource {
+    pub root: PathBuf,
+    pub aliases: Vec<String>,
+    pub mapping: MappingConfig,
+    pub agents: Vec<Agent>,
+    pub skills: Vec<Skill>,
+    pub hooks: Vec<Hook>,
+    pub pi_extensions: Vec<PiExtension>,
+}
+
+impl RefreshSource {
+    pub(crate) fn load(record: &ResolvedSource) -> Self {
+        Self {
+            root: record.root.clone(),
+            aliases: record.aliases.clone(),
+            mapping: MappingConfig::load(&record.root),
+            agents: crate::agent::discover_agents(&record.root.join("agents")).unwrap_or_default(),
+            skills: crate::skill::discover_skills(&record.root.join("skills")).unwrap_or_default(),
+            hooks: crate::hook::discover_hooks(&record.root.join("hooks")).unwrap_or_default(),
+            pi_extensions: crate::pi_extension::discover_pi_extensions(
+                &record.root.join("pi-extensions"),
+            )
+            .unwrap_or_default(),
+        }
+    }
+
+    #[cfg(test)]
+    pub(crate) fn from_root(root: &Path) -> Self {
+        Self::load(&ResolvedSource {
+            root: root.to_path_buf(),
+            aliases: vec![root.to_string_lossy().into_owned()],
+        })
+    }
+}
 
 /// Resolve source directories from lock file entries.
 /// Handles absolute local paths, "." (walks up from CWD), and remote shorthand (cached clones).
 pub(crate) fn resolve_sources(lock: &config::LockFile) -> Vec<PathBuf> {
-    let mut sources: Vec<PathBuf> = Vec::new();
+    resolve_source_records(lock)
+        .into_iter()
+        .map(|source| source.root)
+        .collect()
+}
+
+pub(crate) fn resolve_source_records(lock: &config::LockFile) -> Vec<ResolvedSource> {
+    resolve_source_records_with(lock, resolve_single_source)
+}
+
+fn resolve_source_records_with(
+    lock: &config::LockFile,
+    mut resolver: impl FnMut(&str) -> Option<PathBuf>,
+) -> Vec<ResolvedSource> {
+    let mut sources: Vec<ResolvedSource> = Vec::new();
     let mut seen = std::collections::HashSet::new();
 
     for entry in lock.entries.values() {
-        if seen.contains(&entry.source) {
+        if !seen.insert(entry.source.clone()) {
             continue;
         }
-        seen.insert(entry.source.clone());
-
-        if let Some(dir) = resolve_single_source(&entry.source)
-            && !sources.contains(&dir)
-        {
-            sources.push(dir);
+        if let Some(dir) = resolver(&entry.source) {
+            push_resolved_source(&mut sources, dir, entry.source.clone());
         }
     }
 
-    // Fallback: walk up from CWD to find a vstack source repo
+    // Fallback: walk up from CWD to find a vstack source repo.
     if sources.is_empty()
         && let Ok(mut dir) = std::env::current_dir()
     {
         loop {
             if crate::resolve::is_vstack_source(&dir) {
-                sources.push(dir);
+                let alias = dir.to_string_lossy().into_owned();
+                push_resolved_source(&mut sources, dir, alias);
                 break;
             }
             if !dir.pop() {
@@ -35,21 +93,126 @@ pub(crate) fn resolve_sources(lock: &config::LockFile) -> Vec<PathBuf> {
         }
     }
 
-    // Fallback: try the source registry (cached remote repos)
+    // Fallback: try the source registry (cached remote repos).
     if sources.is_empty() {
         let reg_path = config::source_registry_path();
         if let Ok(registry) = config::SourceRegistry::load(&reg_path) {
             for entry in registry.current.iter().chain(registry.entries.iter()) {
-                if let Some(dir) = resolve_single_source(entry)
-                    && !sources.contains(&dir)
-                {
-                    sources.push(dir);
+                if let Some(dir) = resolver(entry) {
+                    push_resolved_source(&mut sources, dir, entry.clone());
                 }
             }
         }
     }
 
     sources
+}
+
+fn push_resolved_source(sources: &mut Vec<ResolvedSource>, root: PathBuf, alias: String) {
+    if let Some(existing) = sources
+        .iter_mut()
+        .find(|source| same_path(&source.root, &root))
+    {
+        if !existing.aliases.iter().any(|known| known == &alias) {
+            existing.aliases.push(alias);
+        }
+    } else {
+        sources.push(ResolvedSource {
+            root,
+            aliases: vec![alias],
+        });
+    }
+}
+
+pub(crate) fn load_refresh_sources(records: &[ResolvedSource]) -> Vec<RefreshSource> {
+    records.iter().map(RefreshSource::load).collect()
+}
+
+fn canonicalish(path: &Path) -> PathBuf {
+    path.canonicalize().unwrap_or_else(|_| path.to_path_buf())
+}
+
+fn same_path(a: &Path, b: &Path) -> bool {
+    canonicalish(a) == canonicalish(b)
+}
+
+pub(crate) fn refresh_source_for_entry<'a>(
+    sources: &'a [RefreshSource],
+    entry: &config::LockEntry,
+) -> Option<&'a RefreshSource> {
+    if let Some(source) = sources
+        .iter()
+        .find(|source| source.aliases.iter().any(|alias| alias == &entry.source))
+    {
+        return Some(source);
+    }
+
+    let entry_path = Path::new(&entry.source);
+    if entry_path.is_absolute()
+        && let Some(source) = sources
+            .iter()
+            .find(|source| same_path(&source.root, entry_path))
+    {
+        return Some(source);
+    }
+
+    if sources.len() == 1 {
+        sources.first()
+    } else {
+        None
+    }
+}
+
+pub(crate) fn all_source_hooks(sources: &[RefreshSource]) -> Vec<Hook> {
+    sources
+        .iter()
+        .flat_map(|source| source.hooks.iter().cloned())
+        .collect()
+}
+
+pub(crate) fn all_source_pi_extensions(sources: &[RefreshSource]) -> Vec<PiExtension> {
+    sources
+        .iter()
+        .flat_map(|source| source.pi_extensions.iter().cloned())
+        .collect()
+}
+
+pub(crate) fn resolve_skill_pairs_from_sources(
+    names: &[String],
+    lock: &config::LockFile,
+    sources: &[RefreshSource],
+) -> Vec<(String, String)> {
+    names
+        .iter()
+        .map(|name| {
+            let description = lock
+                .entries
+                .get(name)
+                .filter(|entry| entry.kind == ItemKind::Skill)
+                .and_then(|entry| refresh_source_for_entry(sources, entry))
+                .and_then(|source| source.skills.iter().find(|skill| &skill.name == name))
+                .or_else(|| {
+                    sources
+                        .iter()
+                        .flat_map(|source| source.skills.iter())
+                        .find(|skill| &skill.name == name)
+                })
+                .map(|skill| skill.description.clone())
+                .unwrap_or_else(|| name.clone());
+            (name.clone(), description)
+        })
+        .collect()
+}
+
+pub(crate) fn source_pi_extension_for_lock_name<'a>(
+    pi_extensions: &'a [PiExtension],
+    name: &str,
+) -> Option<&'a PiExtension> {
+    pi_extensions.iter().find(|e| e.name == name).or_else(|| {
+        pi_extensions
+            .iter()
+            .find(|e| crate::pi_extension::legacy_names_for(&e.name).contains(&name))
+    })
 }
 
 pub(crate) fn resolve_single_source(source: &str) -> Option<PathBuf> {
@@ -59,7 +222,7 @@ pub(crate) fn resolve_single_source(source: &str) -> Option<PathBuf> {
         return Some(p.to_path_buf());
     }
 
-    // "." — walk up from CWD
+    // "." — walk up from CWD.
     if source == "." {
         let mut dir = std::env::current_dir().ok()?;
         loop {
@@ -73,7 +236,7 @@ pub(crate) fn resolve_single_source(source: &str) -> Option<PathBuf> {
         return None;
     }
 
-    // Remote shorthand (owner/repo) — update and use cached clone
+    // Remote shorthand (owner/repo) — update and use cached clone.
     let cache_dir = config::global_base_dir().join(".vstack").join("cache");
     let key = source.replace('/', "_");
     let cached = cache_dir.join(&key);
@@ -111,6 +274,9 @@ fn update_cached_repo(repo_dir: &std::path::Path) {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::config::{InstallMethod, LockEntry};
+    use std::cell::RefCell;
+    use std::collections::HashMap;
     use std::time::{SystemTime, UNIX_EPOCH};
 
     fn tmpdir(label: &str) -> PathBuf {
@@ -122,6 +288,18 @@ mod tests {
             "vstack-refresh-source-{label}-{}-{nanos}",
             std::process::id()
         ))
+    }
+
+    fn lock_entry(name: &str, source: &str) -> LockEntry {
+        LockEntry {
+            name: name.into(),
+            kind: ItemKind::Agent,
+            source: source.into(),
+            harnesses: vec!["claude-code".into()],
+            method: InstallMethod::Copy,
+            installed_at: "2026-07-03T00:00:00Z".into(),
+            source_hash: String::new(),
+        }
     }
 
     #[test]
@@ -136,6 +314,41 @@ mod tests {
             Some(source.clone())
         );
         assert!(resolve_single_source(&root.to_string_lossy()).is_none());
+
+        let _ = std::fs::remove_dir_all(root);
+    }
+
+    #[test]
+    fn resolve_source_records_calls_resolver_once_per_unique_lock_source() {
+        let root = tmpdir("resolver-count");
+        let source_a = root.join("source-a");
+        let source_b = root.join("source-b");
+        let mut lock = config::LockFile::default();
+        lock.add(lock_entry("rust", "owner/repo"));
+        lock.add(LockEntry {
+            name: "dev".into(),
+            kind: ItemKind::Skill,
+            source: "owner/repo".into(),
+            harnesses: vec!["claude-code".into()],
+            method: InstallMethod::Copy,
+            installed_at: "2026-07-03T00:00:00Z".into(),
+            source_hash: String::new(),
+        });
+        lock.add(lock_entry("scout", "other/repo"));
+
+        let counts: RefCell<HashMap<String, usize>> = RefCell::new(HashMap::new());
+        let records = resolve_source_records_with(&lock, |source| {
+            *counts.borrow_mut().entry(source.to_string()).or_default() += 1;
+            match source {
+                "owner/repo" => Some(source_a.clone()),
+                "other/repo" => Some(source_b.clone()),
+                _ => None,
+            }
+        });
+
+        assert_eq!(records.len(), 2);
+        assert_eq!(counts.borrow().get("owner/repo"), Some(&1));
+        assert_eq!(counts.borrow().get("other/repo"), Some(&1));
 
         let _ = std::fs::remove_dir_all(root);
     }
