+++ cli/src/test_util.rs
@@ -12,6 +12,8 @@ thread_local! {
     static PI_DIR_OVERRIDE: RefCell<Option<PathBuf>> = const { RefCell::new(None) };
     static CODEX_HOME_OVERRIDE: RefCell<Option<PathBuf>> = const { RefCell::new(None) };
     static PROJECT_ROOT_OVERRIDE: RefCell<Option<PathBuf>> = const { RefCell::new(None) };
+    static HOME_DIR_OVERRIDE: RefCell<Option<PathBuf>> = const { RefCell::new(None) };
+    static CONFIG_DIR_OVERRIDE: RefCell<Option<PathBuf>> = const { RefCell::new(None) };
 }
 
 pub(crate) fn pi_dir_override() -> Option<PathBuf> {
@@ -24,6 +26,14 @@ pub(crate) fn codex_home_override() -> Option<PathBuf> {
 
 pub(crate) fn project_root_override() -> Option<PathBuf> {
     PROJECT_ROOT_OVERRIDE.with(|slot| slot.borrow().clone())
+}
+
+pub(crate) fn home_dir_override() -> Option<PathBuf> {
+    HOME_DIR_OVERRIDE.with(|slot| slot.borrow().clone())
+}
+
+pub(crate) fn config_dir_override() -> Option<PathBuf> {
+    CONFIG_DIR_OVERRIDE.with(|slot| slot.borrow().clone())
 }
 
 /// Run `body` with the global Pi dir redirected to `pi_dir` for the current
@@ -42,6 +52,17 @@ pub(crate) fn with_codex_home<R>(codex_home: &Path, body: impl FnOnce() -> R) ->
 /// current test thread, restoring the previous override afterwards.
 pub(crate) fn with_project_root<R>(project_root: &Path, body: impl FnOnce() -> R) -> R {
     with_path_override(&PROJECT_ROOT_OVERRIDE, project_root, body)
+}
+
+/// Run `body` with home/config dirs redirected for the current test thread.
+pub(crate) fn with_home_and_config<R>(
+    home_dir: &Path,
+    config_dir: &Path,
+    body: impl FnOnce() -> R,
+) -> R {
+    with_path_override(&HOME_DIR_OVERRIDE, home_dir, || {
+        with_path_override(&CONFIG_DIR_OVERRIDE, config_dir, body)
+    })
 }
 
 fn with_path_override<R>(
